refactor(macos): extract shared K8s formatters to reduce duplication
- Add K8sFormatters.swift with Optional<String>.k8sAge and Color.podPhase()
- Remove duplicate ageString(from:) from PodListView and ResourceDetailView
- Remove duplicate phase-color functions (phaseColor/podPhaseColor), unify to Color.podPhase()
- PodStatusDot.color now delegates to Color.podPhase() — Succeeded unified to .blue

Fixes SonarCloud duplication threshold violation on PR #33.
